[dispatcher] move to core/ folder, fix imports
